% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_clusters.R
\name{print.SANclusters}
\alias{print.SANclusters}
\title{Print cluster summary}
\usage{
\method{print}{SANclusters}(x, ...)
}
\arguments{
\item{x}{object of class \code{SANclusters} (the result of a call to \code{\link{estimate_clusters}})}

\item{...}{ignored.}
}
\value{
The function prints a summary of the estimated clusters.
}
\description{
Print the cluster-specific sample means and variances of the estimated observational and distributional partition.
}
