% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{print.summary_SANmcmc}
\alias{print.summary_SANmcmc}
\title{Print cluster summary}
\usage{
\method{print}{summary_SANmcmc}(x, ...)
}
\arguments{
\item{x}{object of class \code{SUMmcmc} (the result of a call to \code{\link{summary.SANmcmc}})}

\item{...}{ignored.}
}
\value{
The function prints a summary of the estimated clusters.
}
\description{
Print the cluster-specific sample means and variances of the estimated observational and distributional partition.
}
