% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SANple-package.R
\docType{package}
\name{SANple-package}
\alias{SANple}
\alias{SANple-package}
\title{SANple: Fitting Common and Shared Atoms Nested Models via MCMC}
\description{
An R package for estimating Bayesian nested mixture models via MCMC methods. Specifically, the package implements the common atoms model (Denti et al., 2023), its finite version (D'Angelo et al., 2023), and a hybrid finite-infinite model. All models use Gaussian mixtures with a normal-inverse-gamma prior distribution on the parameters. Additional functions are provided to help analyzing the results of the fitting procedure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/laura-dangelo/SANple}
  \item Report bugs at \url{https://github.com/laura-dangelo/SANple/issues}
}

}
\author{
\strong{Maintainer}: Laura D'Angelo \email{laura.dangelo@live.com} (\href{https://orcid.org/0000-0001-5034-7414}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Francesco Denti \email{francescodenti.personal@gmail.com} (\href{https://orcid.org/0000-0003-2978-4702}{ORCID})
}

}
\keyword{internal}
