% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing.R
\name{summary.SANmcmc}
\alias{summary.SANmcmc}
\alias{plot.SUMmcmc}
\alias{print.SUMmcmc}
\title{Summary of model: estimate observational and distributional clusters}
\usage{
\method{summary}{SANmcmc}(object, burnin = 0, ncores = 1, ...)

\method{plot}{SUMmcmc}(
  x,
  DC_num = NULL,
  type = c("ecdf", "boxplot", "scatter"),
  palette_brewed = FALSE,
  ...
)

\method{print}{SUMmcmc}(x, ...)
}
\arguments{
\item{object}{object of class \code{SANmcmc} (the result of a call to \code{\link{sample_fiSAN}}, 
\code{\link{sample_fSAN}}, or \code{\link{sample_CAM}}).}

\item{burnin}{the length of the burn-in to be discarded before estimating the clusters (default is 2/3 of the iterations).}

\item{ncores}{the number of CPU cores to use, i.e., the number of simultaneous runs at any given time. A value of zero indicates to use all cores on the system.}

\item{...}{ignored.}

\item{x}{an object of class \code{mcmc_clustering}, which can be obtained from the
function \code{\link{summary.SANmcmc}}.}

\item{DC_num}{an integer or a vector of integers indicating which distributional clusters to plot.}

\item{type}{what type of plot should be drawn (only for the left-side plot). Possible types are "boxplot", "ecdf", and "scatter".}

\item{palette_brewed}{(logical) the color palette to be used. Default is \code{R} base colors (\code{palette_brewed = FALSE}).}
}
\value{
a list of class \code{clustering} containing 
\itemize{
  \item \code{obs_level}: a data frame containing the data values, their group indexes, the observational and distributional clustering assignments for each observation.
  \item \code{dis_level}: a vector with the distributional clustering assignment for each unit.
}
}
\description{
Given the MCMC output, estimate the observational and distributional partitions using \code{\link[salso:salso]{salso::salso()}}.
}
\examples{
set.seed(1234)
y <- c(rnorm(400,0,0.3), rnorm(200,5,0.3))
g <- c(rep(1,300), rep(2, 300))
object <- sample_fiSAN(y = y, group = g)
summary(object)

}
\seealso{
\code{\link[salso:salso]{salso::salso()}}, \code{\link{print.SANmcmc}}, \code{\link{plot.SANmcmc}}
}
